Remove randomness from search offset, make it explicit and documented
A mandatory randomness is kinda unreproducible. It is better to leave it up to
the caller if she wants to have a random offset or not. Users who really want to
check all ports >= 1024 for availability might get the response that none are
available just because a few thousands of them have been randomly skipped.