PortUtilTest: use Spock's Spy feature to stub out PortUtil.isAvailable()
Now we do not need to allocate and close a real port anymore in order to
be able to test the case that no port is available in the system. The
trick to limit the search to MAX_PORT_NUMBER is also obsolete now.

Add CGLIB test dependency needed by Spock to mock/spy classes. Otherwise
Spock can only mock/spy interfaces via Java dynamic proxies. For classes
we do need CGLIB.